Couple particle filter with client

+++ pfchangs/app/src/main/java/testclient/filter/Maths.java
@@ -8,7 +8,7 @@ public class Maths {
      * @param y2
      * @return Euclidean distance between (x1, y1) and (x2, y2)
      */
-    public static double distance(float x1, float y1, float x2, float y2) { 
+    public static double distance(double x1, double y1, double x2, double y2) { 
         return Math.sqrt((x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2));
     }
